requirements.txt
pytorch~=1.7.1
torchvision~=0.8.2
pandas~=1.3.5
numpy~=1.19.2
matplotlib~=3.5.1
pillow~=9.2.0
seaborn~=0.11.2
scikit-learn~=1.1.1
ray==1.13.0